Capture periodic stereovision shots for calibration
This code is designed to take 20 snapshots, with a ten second  interval in between each image. This is so that you have time to adjust the position of your checkerboard between each photo. The images will be named according to which camera they came from (left or right) and then numbered according to their order (1-20). The images are then written to a .jpg in the current folder.